# admin-be/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin Backend",
        "description": "Admin Backend Description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://admin-apc.dotbeit.com",
            "description": "Admin Backend"
        }
    ],
    "paths": {
...
        "/api/bookings/{id}": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Booking service",
                "operationId": "Andalusia\\Booking\\Controllers\\API\\BookingController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "booking id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success:": {
                                            "description": "successful action.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error:there is an error.": {
                                            "description": "there is an error.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Booking"
                ],
                "summary": "Booking service",
                "operationId": "Andalusia\\Booking\\Controllers\\API\\BookingController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "booking id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success:": {
                                            "description": "successful action.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error:there is an error.": {
                                            "description": "there is an error.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/notifications/{id}": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Notification service",
                "operationId": "Andalusia\\Notification\\Controllers\\API\\MainController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "notification id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success:": {
                                            "description": "successful action.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error:there is an error.": {
                                            "description": "there is an error.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Notification"
                ],
                "summary": "Notification service",
                "operationId": "Andalusia\\Notification\\Controllers\\API\\MainController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "notification id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "image": {
                                        "description": "an image",
                                        "type": "file",
                                        "example": "image.png"
                                    },
                                    "thumbnail": {
                                        "description": "a thumbnail",
                                        "type": "file",
                                        "example": "thumbnail.png"
                                    },
                                    "link": {
                                        "description": "the link to the notification",
                                        "type": "string",
                                        "example": "http://www.andalusiagroup.net/"
                                    },
                                    "message": {
                                        "description": "the message that will be displayed to the user",
                                        "type": "string",
                                        "example": "special offer for you!"
                                    },
                                    "type": {
                                        "description": "is this notification is just a normal notification or it's an offer notification?",
                                        "type": "string",
                                        "enum": [
                                            "normal",
                                            "offer"
                                        ],
                                        "example": "offer"
                                    },
                                    "is_read": {
                                        "description": "mark the notification as read ... note that: to mark the notification as read you have to send this property only and don't include other properties",
                                        "type": "boolean",
                                        "enum": [
                                            1,
                                            true
                                        ],
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success:": {
                                            "description": "successful action.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error:there is an error.": {
                                            "description": "there is an error.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notification"
                ],
                "summary": "Notification service",
                "operationId": "Andalusia\\Notification\\Controllers\\API\\MainController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "notification id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success:": {
                                            "description": "successful action.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error:there is an error.": {
                                            "description": "there is an error.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/surveys_answers/{id}": {
            "get": {
                "tags": [
                    "Survey Answer"
                ],
                "summary": "Survey Question Answers service",
                "operationId": "Andalusia\\Survey\\Controllers\\API\\AnswerController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "answer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success:": {
                                            "description": "successful action.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error:there is an error.": {
                                            "description": "there is an error.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Survey Answer"
                ],
                "summary": "Survey Question Answers service",
                "operationId": "Andalusia\\Survey\\Controllers\\API\\AnswerController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "answer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success:": {
                                            "description": "successful action.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error:there is an error.": {
                                            "description": "there is an error.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/surveys_questions/{id}": {
            "get": {
                "tags": [
                    "Survey Question"
                ],
                "summary": "Survey Question service",
                "operationId": "Andalusia\\Survey\\Controllers\\API\\QuestionController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "question id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success:": {
                                            "description": "successful action.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error:there is an error.": {
                                            "description": "there is an error.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Survey Question"
                ],
                "summary": "Survey Question service",
                "operationId": "Andalusia\\Survey\\Controllers\\API\\QuestionController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "question id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success:": {
                                            "description": "successful action.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error:there is an error.": {
                                            "description": "there is an error.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/surveys_users_answers/{id}": {
            "get": {
                "tags": [
                    "Survey User Answer"
                ],
                "summary": "Survey User Answer service",
                "operationId": "Andalusia\\Survey\\Controllers\\API\\UserAnswerController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user answer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success:": {
                                            "description": "successful action.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error:there is an error.": {
                                            "description": "there is an error.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Survey User Answer"
                ],
                "summary": "Survey User Answer service",
                "operationId": "Andalusia\\Survey\\Controllers\\API\\UserAnswerController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user answer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success:": {
                                            "description": "successful action.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error:there is an error.": {
                                            "description": "there is an error.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}